+++ README.md
@@ -3,7 +3,7 @@
 [![Build Status](https://travis-ci.org/GitbookIO/filterable.png?branch=master)](https://travis-ci.org/GitbookIO/filterable)
 [![NPM version](https://badge.fury.io/js/filterable.svg)](http://badge.fury.io/js/filterable)
 
-Filterable is a Node.JS module to parse [GitHub-like search queries](https://help.github.com/articles/search-syntax/), for example: `cats stars:>10`. It can generates MongoDB queries.
+Filterable is a Node.JS module to parse [GitHub-like search queries](https://help.github.com/articles/search-syntax/), for example: `cats stars:>10`. It can generates MongoDB or Elasticsearch queries.
 
 This module is perfect to integrates complex search (like GitHub) into your application. It integrates perfectly with MongoDB but can also easily be adapted to other databases.
 
@@ -28,127 +28,32 @@ Install it using:
 $ npm install filterable
 ```
 
-Create a filterable type of object:
+Parse a query string:
 
 ```js
 var filterable = require("filterable");
 
-var filterable = new filterable.Filterable({
-    // Fields containing search tags (default is "tags")
-    tags: "tags",
-
-    // Fields to use for indexation, extend the Strign fields
-    tagsFields: [ "description" ],
-
-    // List of fields that are authorized
-    fields: {
-        name: {
-            type: String
-        },
-        mail: {
-            type: String,
-
-            // In the MongoDB query, the field will be "email"
-            alias: "email"
-        },
-        username: {
-            type: String
-        },
-        followers: {
-            type: Number
-        },
-        language: {
-            type: String,
-
-            // Alias could also be an array
-            alias: ["language_detected", "language_settings"]
-        },
-        friends: {
-            type: String,
-
-            // Transform query value, 
-            value: function(value, next) {
-                getUserRef(value, next);
-            }
-        }
-    }
-});
+var query = filterable.Query('cats stars:>10')
+    .parse();
 ```
 
-#### Queries
-
-Generate a mongo query from a string:
+Output query for a database:
 
 ```js
-filterable.query("mail:samypesse@gmail.com", function(err, query) {
-    /*
-    query.toMongo() == {
-        "email": {
-            "$eq": "samypesse@gmail.com"
-        }
-    }
-    */
-});
-
+var mongQuery = query.toMongo();
+var elasticQuery = query.toElasticsearch();
 ```
+
+Filter and customize queries using `QueryBuilder`:
 
 ```js
-filterable.query("cats followers:>10", function(err, query) {
-    /*
-    query.toMongo() == {{
-        "tags": {
-            "$in": ["cats"]
-        },
-        "followers": {
-            "$gt": 10
-        }
-    }
-    */
-});
-```
-
-#### Tags
-
-Generate list of tags from a text content. You'll just need to index this array of string into the object.
-
-```js
-filterable.tags("Hello World");
-
-["hello", "world"]
-```
-
-Generate list of tags using searchable fields:
-
-```js
-filterable.tags({
-    name: "Samy",
-    username: "SamyPesse",
-    description: "This is my profile"
+var builder = filterable.QueryBuilder({
+    // Field using when quering with text
+    textField: 'description'
 });
 
-["samy", "samypesse", "this", "is", "my", "profile"]
+
+// Parse queries (return a Query object)
+var q1 = builder.parse('Hello world')
+var q2 = builder.parse('cats stars:>10');
 ```
-
-#### Mongoose
-
-This module contains a plugin for [mongoose](https://github.com/LearnBoost/mongoose):
-
-```js
-userSchema.plugin(filterable.mongoose, {
-    tags: "tags",
-    fields: ["username", "name", "location"],
-    tagsFields: [ "description" ]
-    alias: {
-        "username": "userId"
-    }
-});
-```
-
-You can then run search query:
-
-```js
-User.search("username:test").count().exec(function(err, n) {
-    
-});
-```
-
