+++ package.json
@@ -11,14 +11,13 @@
         "url": "https://github.com/GitbookIO/filterable/issues"
     },
     "engines": {
-        "node": ">=0.6"
+        "node": ">=0.12"
     },
     "dependencies": {
-        "lodash": "2.4.1",
-        "q": "1.1.2"
+        "lodash": "^3.10.1"
     },
     "devDependencies": {
-        "mocha": "1.18.2",
+        "mocha": "2.3.4",
         "object-hash": "0.4.0"
     },
     "repository": {
@@ -29,6 +28,6 @@
         "test": "mocha --reporter list"
     },
     "keywords": [
-        "github", "mongoose", "mongodb", "search", "filter", "query"
+        "github", "mongoose", "mongodb", "search", "filter", "query", "elasticsearch"
     ]
 }
